requirements.txt
blinker==1.9.0                  # Señales para aplicaciones Flask (usado internamente por Flask)
certifi==2025.4.26              # Certificados raíz para validar conexiones SSL
charset-normalizer==3.4.2       # Detección y normalización de codificación de texto (usado por requests)
click==8.2.1                    # Utilidad para crear interfaces de línea de comandos (usado por Flask)
colorama==0.4.6                 # Colorea la salida en terminal (útil en Windows)
Flask==3.1.1                    # Framework web principal del proyecto
Flask-MySQL==1.6.0              # Extensión para conectar Flask con bases de datos MySQL
Flask-SQLAlchemy==3.1.1         # Integración de SQLAlchemy (ORM) con Flask
greenlet==3.2.2                 # Soporte para concurrencia ligera (usado por SQLAlchemy)
idna==3.10                      # Soporte para nombres de dominio internacionalizados (usado por requests)
itsdangerous==2.2.0             # Firmado seguro de datos (usado por Flask para sesiones y tokens)
Jinja2==3.1.6                   # Motor de plantillas usado por Flask
MarkupSafe==3.0.2               # Escapado seguro de HTML (usado por Jinja2)
mysql-connector-python==9.4.0   # Conector oficial de MySQL para Python
packaging==25.0                 # Utilidades para manejo de versiones y dependencias
pillow==11.3.0                  # Procesamiento de imágenes en Python
PyMySQL==1.1.1                  # Otro conector MySQL puro en Python
pytesseract==0.3.13             # Wrapper para Tesseract OCR (reconocimiento óptico de caracteres)
python-dotenv==1.1.0            # Carga variables de entorno desde archivos .env
requests==2.32.3                # Cliente HTTP para consumir APIs y recursos web
SQLAlchemy==2.0.41              # ORM para bases de datos relacionales
typing_extensions==4.13.2       # Tipos adicionales para anotaciones de tipo en Python
urllib3==2.4.0                  # Biblioteca HTTP de bajo nivel (usada por requests)
Werkzeug==3.1.3                 # Utilidades WSGI, servidor y seguridad (usado por Flask)
